+++ soccer/src/soccer/planning/planner/plan_request.hpp
@@ -25,75 +25,20 @@ namespace planning {
  * robot path to be planned.
  */
 struct PlanRequest {
-    PlanRequest(RobotInstant start, MotionCommand command,  // NOLINT
-                RobotConstraints constraints, rj_geometry::ShapeSet field_obstacles,
-                rj_geometry::ShapeSet virtual_obstacles, TrajectoryCollection* planned_trajectories,
-                unsigned shell_id, const WorldState* world_state, int8_t priority = 0,
-                rj_drawing::RosDebugDrawer* debug_drawer = nullptr, bool ball_sense = false,
-                float min_dist_from_ball = 0, float dribbler_speed = 0)
+    PlanRequest(RobotInstant start, RobotIntent robotIntent,  // NOLINT
+                RobotConstraints constraints,
+                 TrajectoryCollection* planned_trajectories,
+                 int8_t priority = 0,
+                rj_drawing::RosDebugDrawer* debug_drawer = nullptr,
+                 float dribbler_speed = 0)
         : start(start),
-          motion_command(command),  // NOLINT
           constraints(constraints),
-          field_obstacles(std::move(field_obstacles)),
-          virtual_obstacles(std::move(virtual_obstacles)),
           planned_trajectories(planned_trajectories),
-          shell_id(shell_id),
           priority(priority),
-          world_state(world_state),
           debug_drawer(debug_drawer),
-          ball_sense(ball_sense),
-          min_dist_from_ball(min_dist_from_ball),
           dribbler_speed(dribbler_speed) {}
 
-    /**
-     * The robot's starting state.
-     */
-    RobotInstant start;
-
-    /**
-     * The goal to plan for.
-     */
-    MotionCommand motion_command;
-
-    /**
-     * Angular and linear acceleration and velocity constraints on the robot.
-     */
-    RobotConstraints constraints;
-
-    /**
-     * The list of field obstacles.
-     */
-    rj_geometry::ShapeSet field_obstacles;
-
-    /**
-     * The list of "virtual" obstacles, generated by gameplay representing soft
-     * constraints.
-     */
-    rj_geometry::ShapeSet virtual_obstacles;
-
-    /**
-     * Trajectories for each of the robots that has already been planned.
-     * nullptr for unplanned robots.
-     */
-    TrajectoryCollection* planned_trajectories;
-
-    /**
-     * The robot's shell ID. Used for debug drawing.
-     */
-    unsigned shell_id;
-
-    /**
-     * The state of the world, containing robot and ball states.
-     *
-     * For obstacle-avoidance purposes, obstacles should be used instead. This
-     * can be used for lookup of robots/balls by ID.
-     */
-    const WorldState* world_state;
-
-    /**
-     * The priority of this plan request.
-     */
-    int8_t priority;
+    GlobalState& global_state;
 
     /**
      * Allows debug drawing in the world. If this is nullptr, no debug drawing
@@ -101,18 +46,6 @@ struct PlanRequest {
      */
     rj_drawing::RosDebugDrawer* debug_drawer;
 
-    // Whether the robot has a ball
-    bool ball_sense = false;
-
-    /**
-     * How far away to stay from the ball, if the MotionCommand chooses to avoid the ball.
-     */
-    float min_dist_from_ball = 0;
-
-    /**
-     * Dribbler Speed
-     */
-    float dribbler_speed = 0;
 };
 
 /**
